Add more UI updates
- Border updates
- Padding and spacing updates
- Added "fade" effect to anouncement section to cover text behind
  ellipses.

The "fade" affect works by inserting a gradient that gradually goes from
fully transparent on the left to fully opaque on the right (to the
background color), and is set to position absolute to stay at the
bottom. Initially, there were situations where the fade would sit
perfecty over the text, then other zoom levels would make it sit halfway
over, so I needed to implement steps for the height to grow by so that
the fade would either sit straight over the text or not at all. This was
implemented with display: grid on the parent container, using
repeat(auto-fill, 20px) (20px is the size of the font) to automatically
generate new rows in space that is available. Then, the text spans from
the second to the last row, and the fade is on the last row as well.
The grid is an elegant, non-hacky solution to adding in the text fade to
show the user that more text is available or not without breaking on
window zooming.